webinar-cloudflare-worker/pyproject.toml
[project]
name = "webinar-backend"
dependencies = [
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "motor>=3.5.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.5.0",
]

